+++ packages/@lwc/engine/src/framework/base-bridge-element.ts
@@ -13,9 +13,12 @@ import {
     create,
     getOwnPropertyNames,
     isUndefined,
-    defineProperties,
     seal,
     ArraySlice,
+    setPrototypeOf,
+    defineProperty,
+    isFunction,
+    defineProperties,
 } from "../shared/language";
 import { getCustomElementVM } from "./vm";
 import {
@@ -83,7 +86,33 @@ export interface HTMLIFrameElementConstructor {
 }
 
 export function HTMLBridgeElementFactory(SuperClass: HTMLElementConstructor, props: string[], methods: string[]): HTMLElementConstructor {
-    class HTMLBridgeElement extends SuperClass {}
+    let HTMLBridgeElement;
+    /**
+     * Modern browsers will have all Native Constructors as regular Classes
+     * and must be instantiated with the new keyword. In older browsers,
+     * specifically IE11, those are objects with a prototype property defined,
+     * since they are not supposed to be extended or instantiated with the
+     * new keyword. This forking logic supports both cases, specifically because
+     * wc.ts relies on the construction path of the bridges to create new
+     * fully qualifying web components.
+     */
+    if (isFunction(SuperClass)) {
+        HTMLBridgeElement = class extends SuperClass {};
+    } else {
+        HTMLBridgeElement = function() {
+            // Bridge classes are not supposed to be instantiated directly in
+            // browsers that do not support web components.
+            throw new TypeError('Illegal constructor');
+        };
+        // prototype inheritance dance
+        setPrototypeOf(HTMLBridgeElement, SuperClass);
+        setPrototypeOf(HTMLBridgeElement.prototype, SuperClass!.prototype);
+        defineProperty(HTMLBridgeElement.prototype, 'constructor', {
+            writable: true,
+            configurable: true,
+            value: HTMLBridgeElement,
+        });
+    }
     const descriptors: PropertyDescriptorMap = create(null);
     // expose getters and setters for each public props on the new Element Bridge
     for (let i = 0, len = props.length; i < len; i += 1) {
@@ -105,7 +134,7 @@ export function HTMLBridgeElementFactory(SuperClass: HTMLElementConstructor, pro
         };
     }
     defineProperties(HTMLBridgeElement.prototype, descriptors);
-    return HTMLBridgeElement;
+    return HTMLBridgeElement as HTMLElementConstructor;
 }
 
 export const BaseBridgeElement = HTMLBridgeElementFactory(HTMLElement, getOwnPropertyNames(HTMLElementOriginalDescriptors), []);